requirements.txt
flet>=0.80.0
langgraph>=0.2.0
google-genai>=0.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
pyperclip>=1.8.0
pyinstaller>=6.0.0

